requirements.txt
numpy>=1.26.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.2.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
transformers>=4.35.0
huggingface-hub>=0.20.0
safetensors>=0.4.0
requests>=2.31.0
llama-cpp-python>=0.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pytest>=7.4.0
